+++ Android/app/src/main/java/com/example/boardgamesocial/MainActivity.java
@@ -38,7 +38,6 @@ public class MainActivity extends AppCompatActivity {
     private EditText etUsername, etPassword;
     private Button btnLogin, btnSignup;
 
-    private RetrofitClient retrofitClient;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -49,34 +48,6 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void wireUpDisplay() {
-        retrofitClient = RetrofitClient.getClient();
-
-        // This token needs to be acquired with login request
-        HeaderInterceptor.token.setToken("e4a36ccc86bc71b7e78c5d42bbd3109ab4764af1");
-        retrofitClient.getCall(Game.class, new HashMap<String, String>(){{
-            put("gameTitle", "7 Wonders Duel");
-        }}).enqueue(new Callback<JsonArray>() {
-            @Override
-            public void onResponse(Call<JsonArray> call, Response<JsonArray> response) {
-                if (response.isSuccessful()) {
-                    assert response.body() != null;
-                    List<Game> list = getObjectList(response.body(), Game.class);
-                    Log.i("Game", list.get(0).getGameTitle());
-                } else {
-                    new Exception("Request failed, code: " + response.code()).printStackTrace();
-                }
-            }
-
-            @Override
-            public void onFailure(Call<JsonArray> call, Throwable t) {
-                try {
-                    throw t;
-                } catch (Throwable throwable) {
-                    throwable.printStackTrace();
-                }
-            }
-        });
-
         etUsername = findViewById(R.id.et_login_username);
         etPassword = findViewById(R.id.et_login_password);
         btnLogin = findViewById(R.id.btn_login);
